Refactors into 3 components
