+++ client/src/views/App/Home2.js
@@ -7,27 +7,8 @@ import { NavigateBefore } from "styled-icons/material/NavigateBefore";
 import { NavigateNext } from "styled-icons/material/NavigateNext";
 import { Compare } from "styled-icons/material/Compare";
 import { connect } from "react-redux";
-
-const BlackLeft = styled(NavigateBefore)`
-  color: grey;
-  height: 35px;
-  width: 35px;
-  border-radius: 6px;
-  :hover {
-    background: rgb(224, 224, 224);
-  }
-  cursor: pointer;
-`;
-const BlackRight = styled(NavigateNext)`
-  color: black;
-  height: 35px;
-  width: 35px;
-  border-radius: 6px;
-  :hover {
-    background: rgb(224, 224, 224);
-  }
-  cursor: pointer;
-`;
+import Charts from "../../components/HomeView/Charts.js";
+import Compare2 from "../../components/HomeView/Compare/Compare2.js";
 
 const BlackCompare = styled(Compare)`
   color: black;
@@ -51,19 +32,10 @@ class Home2 extends Component {
           </div>
           <div className="right-pane">
             <div className="right-top">
-              <header>
-                <h2>Charts</h2>
-                <div className="toggle">
-                  <BlackLeft />
-                  <BlackRight />
-                </div>
-              </header>
+              <Charts />
             </div>
             <div className="right-bottom">
-              <header>
-                <h2>Compare</h2>
-                <BlackCompare />
-              </header>
+              <Compare2 />
             </div>
           </div>
         </div>
